Invert `qiskit-quantum-info` and `qiskit-circuit` dependency
This moves the last remaining dependency from `qiskit-quantum-info` out
of that package and into `qiskit-circuit` itself; rather than
`VersorU2` consuming the `StandardGate` to construct itself, we instead
have the `StandardGate` learn how to construct `VersorU2` from itself.

This allows us to invert the dependency, and put `qiskit-quantum-info`
below `qiskit-circuit` in the crate hierarchy, opening up the ability to
use `qiskit-quantum-info` types natively within core circuit IR types,
and to re-use `quantum-info` routines around conversion of matrix-free
objects to matrix form, and so on.
