+++ crates/quantum_info/Cargo.toml
@@ -10,7 +10,6 @@ name = "qiskit_quantum_info"
 doctest = false
 
 [dependencies]
-qiskit-circuit.workspace = true
 qiskit-util = { workspace = true, features = ["py"] }
 numpy.workspace = true
 num-complex.workspace = true
@@ -50,3 +49,6 @@ features = ["hashbrown", "indexmap", "num-complex", "num-bigint", "smallvec"]
 
 [lints]
 workspace = true
+
+[dev-dependencies]
+qiskit-circuit.workspace = true
